Add tests on all class methods
